% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_adae.R
\name{mk_adae_demo}
\alias{mk_adae_demo}
\title{Demo function to make ADAE data for validation purposes}
\usage{
mk_adae_demo(study_metadata)
}
\arguments{
\item{study_metadata}{}
}
\value{
data.table
}
\description{
Demo function to make ADAE data for validation purposes
}
