% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramnog-package.R
\docType{package}
\name{ramnog-package}
\alias{ramnog}
\alias{ramnog-package}
\title{ramnog: Documentation Hub for the 'chef' Ecosystem}
\description{
ramnog provides vignettes and documentation for the chef ecosystem of
R packages for clinical trial evidence generation following AMNOG guidelines.
The core functionality is provided by \pkg{chef}, \pkg{chefStats}, and
\pkg{chefCriteria}.
}
\section{Vignettes}{

Use \code{browseVignettes("ramnog")} to see all available vignettes.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://hta-pharma.github.io/ramnog/}
  \item \url{https://github.com/hta-pharma/ramnog}
  \item Report bugs at \url{https://github.com/hta-pharma/ramnog/issues}
}

}
\author{
\strong{Maintainer}: Christian Haargaard Olsen \email{cino@novonordisk.com}

Authors:
\itemize{
  \item Matthew David Phelps \email{mewp@novonordisk.com}
  \item Nicolai Skov Johnsen \email{nosj@novonordisk.com}
  \item Henrik Sparre Spiegelhauer \email{hspu@novonordisk.com}
}

}
\keyword{internal}
